+++ src/auth.ts
@@ -1,9 +1,27 @@
+/**
+ * This file contains the auth functions.
+ *
+ * @file src\auth.ts
+ * @author Luca Liguori
+ * @date 2024-05-01
+ * @version 1.0.0
+ *
+ * Copyright 2024, 2025 Luca Liguori.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License. *
+ */
+
 import crypto from 'crypto';
-/*
-import { AnsiLogger, TimestampFormat, db, idn, rs, zb } from 'node-ansi-logger';
-import { Shelly } from './shelly.js';
-import { ShellyDevice } from './shellyDevice.js';
-*/
 
 export interface AuthParams {
   realm: string; // device_id
@@ -53,6 +71,22 @@ export function getGen1BodyOptions(params?: Record<string, string | number | boo
   return new URLSearchParams(params as any).toString();
 }
 
+/*
+options: {
+  "method":"POST",
+  "headers": {
+    "Content-Type":"application/json"
+  },
+  "body": "{ 
+    \"jsonrpc\":\"2.0\",
+    \"id\":10,
+    \"src\":\"Matterbridge\",
+    \"method\":\"Cover.GoToPosition \",
+    \"params\":{\"id\":0,\"pos\":90}
+  }"
+}
+*/
+
 export function getGen2BodyOptions(jsonrpc: string, id: number, src: string, method: string, params?: Record<string, string | number | boolean>, auth?: AuthParams): string {
   const body: Record<string, string | number | boolean | object | AuthParams> = {};
   body.jsonrpc = '2.0';
@@ -61,6 +95,7 @@ export function getGen2BodyOptions(jsonrpc: string, id: number, src: string, met
   body.method = method;
   if (params) body.params = params;
   if (auth) body.auth = auth;
+  // console.log(JSON.stringify(body));
   return JSON.stringify(body);
 }